deleted cart.js from api directory
